+++ include/types.h
@@ -4,8 +4,8 @@
 /** 文件迭代器 */
 typedef struct FileIteratorRec_* FileIterator;
 typedef struct FileIteratorRec_ {
-	unsigned int index;		/**< 当前索引位置 */
-	unsigned int length;		/**< 文件列表总长度 */
+	size_t index;			/**< 当前索引位置 */
+	size_t length;			/**< 文件列表总长度 */
 	void *privdata;			/**< 私有数据 */
 	char *filepath;			/**< 文件路径 */
 	void(*unlink)(FileIterator);	/**< 将当前文件从列表中移除 */
